+++ lib/events_api_web/controllers/auth/error_handler.ex
@@ -1,0 +1,20 @@
+defmodule EventsApiWeb.Auth.ErrorHandler do
+  @moduledoc """
+  this model has functions that return guardian errors
+  """
+  use EventsApiWeb, :controller
+
+  def auth_error(conn, {type, _reason}, _opts) do
+    case type do
+      :invalid_token ->
+        conn
+        |> put_resp_content_type("aplication/json")
+        |> send_resp(401, Jason.encode!(%{error: "Token expirado ou inválido"}))
+
+      :unauthenticated ->
+        conn
+        |> put_resp_content_type("aplication/json")
+        |> send_resp(401, Jason.encode!(%{error: "Token não encontrado"}))
+    end
+  end
+end