+++ lib/events_api_web/controllers/event_controller.ex
@@ -12,6 +12,10 @@ defmodule EventsApiWeb.EventController do
   end
 
   def create(conn, %{"event" => event_params}) do
+    user = Guardian.Plug.current_resource(conn)
+
+    event_params = Map.put(event_params, "user_id", user.id)
+
     with {:ok, event} <- Doings.create_event(event_params) do
       conn
       |> put_status(:created)