+++ lib/events_api_web/controllers/user_controller.ex
@@ -2,6 +2,7 @@ defmodule EventsApiWeb.UserController do
   use EventsApiWeb, :controller
 
   alias EventsApi.Accounts
+  alias EventsApiWeb.Auth.Guardian
 
   action_fallback EventsApiWeb.FallbackController
 
@@ -11,6 +12,14 @@ defmodule EventsApiWeb.UserController do
       |> put_status(:created)
       |> put_resp_header("location", Routes.user_path(conn, :show, user))
       |> render("show.json", user: user)
+    end
+  end
+
+  def signin(conn, %{"email" => email, "password" => password}) do
+    with {:ok, user, token} <- Guardian.authenticate(email, password) do
+      conn
+      |> put_status(:created)
+      |> render("user_auth.json", user: user, token: token)
     end
   end